Deduplicated SearchMessagForm related code
